# docs/swagger.yaml
definitions:
...
  entity.ProductWithStockAndWarehouse:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      product_id:
        type: integer
      status:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
      warehouse_id:
        type: integer
    type: object
  entity.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve all products from the database
      produces:
      - application/json
      responses:
        "200":
          description: Product getted
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetProducts error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Retrieve all products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Add multiple products to the database
      parameters:
      - description: Array of products to add
        in: body
        name: Object
        required: true
        schema:
          items:
            $ref: '#/definitions/entity.ProductWithStockAndWarehouse'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Products added
          schema:
            $ref: '#/definitions/entity.Response'
        "422":
          description: Parse input error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Add products
      tags:
      - Product
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product from the database by ID
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid product_id
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application DeleteProduct error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Delete a product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Retrieve a product from the database by ID
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product retrieved
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid product_id
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetProduct error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Retrieve a product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Update a product in the database by ID
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Product updated
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid product_id
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application UpdateProduct error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Update a product
      tags:
      - Product
  /products/search:
    get:
      consumes:
      - application/json
      description: Search for products by keyword
      parameters:
      - description: Search keyword for product name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Products searched
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application SearchProducts error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Search for products
      tags:
      - Product
swagger: "2.0"